+++ src/main/java/com/avg/lawsuitmanagement/token/repository/TokenMapperRepository.java
@@ -1,17 +1,13 @@
 package com.avg.lawsuitmanagement.token.repository;
 
 
-import com.avg.lawsuitmanagement.token.dto.MemberDto;
 import com.avg.lawsuitmanagement.token.dto.RefreshTokenDto;
 import com.avg.lawsuitmanagement.token.repository.param.RefreshTokenParam;
-import com.avg.lawsuitmanagement.token.repository.param.SignUpParam;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface TokenMapperRepository {
 
-    MemberDto selectMemberByEmail(String email);
-    void insertMember(SignUpParam param);
     void insertRefreshToken(RefreshTokenParam param);
     RefreshTokenDto selectRefreshTokenByKey(String key);
 }
