+++ src/main/resources/mybatis/mapper/Member.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.avg.lawsuitmanagement.member.repository.MemberMapperRepository">
+
+    <select id="selectMemberByEmail" parameterType="java.lang.String">
+        select m.*, h.name as hierarchy, r.name as role
+        from member m
+                 join hierarchy h on m.hierarchy_id = h.id
+                 join role r on m.role_id = r.id
+        where m.email = #{email}
+          and m.is_deleted = false
+    </select>
+
+    <insert id="insertMember" parameterType="SignUpParam">
+        insert into member (email, password, name, phone, hierarchy_id, address, role_id)
+        values (#{email}, #{password}, #{name}, #{phone}, #{hierarchyId}, #{address}, #{roleId})
+    </insert>
+</mapper>
