--- src/main/resources/mybatis/mapper/Sample.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<!-- 위 코드가 mapper 설정을 하는 코드 -->
-
-<!-- Mapper를 사용하는 Repository의 전체 경로를 적어줘야함. -->
-<mapper namespace="com.avg.lawsuitmanagement.sample.SampleMapperRepository">
-
-    <insert id="save" parameterType="SampleDto">
-        insert into sample(data)
-        values (#{data})
-    </insert>
-
-    <select id = "selectAll" resultType="SampleDto">
-        select * from sample
-    </select>
-
-    <select id = "selectSampleByData" parameterType="java.lang.String" resultType="SampleDto">
-        select * from sample where data = #{data}
-    </select>
-
-</mapper>
