+++ src/main/resources/mybatis/mapper/Token.xml
@@ -5,20 +5,6 @@
 
 <mapper namespace="com.avg.lawsuitmanagement.token.repository.TokenMapperRepository">
 
-    <select id="selectMemberByEmail" parameterType="java.lang.String">
-        select m.*, h.name as hierarchy, r.name as role
-        from member m
-                 join hierarchy h on m.hierarchy_id = h.id
-                 join role r on m.role_id = r.id
-        where m.email = #{email}
-          and m.is_deleted = false
-    </select>
-
-    <insert id="insertMember" parameterType="SignUpParam">
-        insert into member (email, password, name, phone, hierarchy_id, address, role_id)
-        values (#{email}, #{password}, #{name}, #{phone}, #{hierarchyId}, #{address}, #{roleId})
-    </insert>
-
     <insert id="insertRefreshToken" parameterType="RefreshTokenParam">
         replace into refresh_token (`key`, `value`)
         values (#{key}, #{value})
